requirements.txt
asgiref==3.2.7
Django==3.0.6
django-debug-toolbar==2.2
django-shortuuidfield==0.1.3
djangorestframework==3.11.0
Pillow==7.1.2
PyMySQL==0.9.3
pytz==2020.1
shortuuid==1.0.1
six==1.14.0
sqlparse==0.3.1
